requirements.txt
jetpack-io==0.5.1.dev202203301648638711
Pillow==9.0.1
boto3==1.21.29